mon test dist: Allow confgen with no pollers
Specifying --pgroups=0 would previously create a peer group without
nodes in it, which in turn would cause us to bail out pretty hard
when creating config for it. With this patch, we simply skip groups
with 0 nodes and we can once again use "mon test dist" for generating
a single-instance config.

Signed-off-by: Andreas Ericsson <ae@op5.se>