mon test dist: Catch exceptions in fake_mesh.ptest()
Sometimes we run tests that throw exceptions until everything is
properly initialized, which causes the entire testsuite to bail out
with a hard error, when the tests would have succeeded had they just
gotten a few more seconds to run.

This patch provides those extra seconds, so long as we don't run out
of time.

Signed-off-by: Andreas Ericsson <ae@op5.se>