logging: Don't always write to stderr where the log goes
I really don't care, and it's noisy.

Because we haven't parsed the config yet, using the configured log level
to determine if we should print where the log goes would be interesting.
But there is a debug flag that gets passed to merlin as a CLI parameter
- let's hook into that. If you tell nagios to go into debug mode, it'll
still tell you where the log goes. Otherwise, it will assume you know
what you wrote in your config.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>