mon log sortmerge: Fix argvec creation for the import helper
After having run the sortmerge thing, we're supposed to run the
import helper, but previously we passed it a list made up of a few
strings and one list. Since we can only pass strings as elements in
the list, that didn't pan out so well. This converts it so we concat
the arg list to the list we create, which makes things work out just
fine again.

Signed-off-by: Andreas Ericsson <ae@op5.se>