ocimp: Use the proper variable when parsing cfgfile compound variables
"vlist_len" is internal and marks the size of the variable list.
"vars" is external and marks the number of variables available in
that same list.

It's important not to confuse the two, as it leads to segfaults when
parsing the key/values out of compounds, as happened before when we
ran into any kind of downtime that happened to have anything but
exactly 15 variables (15 because it ended up on an ALLOC_GROW()
boundary, and downtimes always have either 14 or 15).

Signed-off-by: Andreas Ericsson <ae@op5.se>