mon start: Fix synergy config removal hack
If "mon start" is run on a poller node, it has tried to run a sed
command to remove syngery's cfg_dir entry in nagios.cfg.

However, the sed expression was broken and the file was never modified,
except the file's mtime. Changing the mtime of nagios.cfg all the time
is not awesome.

This fixes the sed expression to actually remove the cfg_dir entry, and
it only does this if the synergy cfg_dir entry actually exists in the
file. In other words, no more mtime updates, and bug #7771 = fixed.

Signed-off-by: Philip Eklöf <peklof@op5.com>