mon syscheck: Add initial system check plugins
This adds several bash scripts to the "mon sysconf check"
framework. This is just a very basic set of all the many
things that are possible to check.

Some of the modules are *very* static, such as the proc_*
ones. They expect to find exactly matching command lines.
Just a single argument added to the command line of one of
these processes would make the checks go CRITICAL. But that
is a good thing - we want to find deviating settings etc;
it helps up troubleshoot non-working systems.

The staticness also makes it very easy to add new checks,
but also to keep the check plugins argumentless.
Differences between Monitor versions etc. needs to be
handled in the future, though.

Descriptions fetched from the "mon syscheck" help text:

 db_mysql_check
   Detects errors in the merlin and nacoma database using mysqlcheck.

 fs_free_inodes
   Checks all locally mounted filesystems for low amounts of free inodes.

 fs_free_space
   Checks all locally mounted filesystems for low amounts of free space.

 fs_state
   Checks for badly mounted filesystems and non-clean ext filesystems.

 license
   Verifies that the op5 license file is installed and valid.

 proc_merlind [--info]
   Checks the running state of process *merlind*

 proc_monitor_core [--info]
   Checks the running state of process *monitor core*

 proc_monitor_worker [--info]
   Checks the running state of process *monitor worker*

 proc_mysqld [--info]
   Checks the running state of process *mysqld*

 proc_npcd [--info]
   Checks the running state of process *npcd*

 proc_op5kad [--info]
   Checks the running state of process *op5kad*

 proc_rrdcached [--info]
   Checks the running state of process *rrdcached*

Signed-off-by: Philip Eklöf <peklof@op5.com>