install: refactor database update & installation
This patch makes database update somewhat more reasonable,
by adding a set_initial_db_version procedure which does what you'd
think (it sets the initial DB version, if no version is set). This way,
we can avoid dropping the version table on each update, which allows us
to not have to keep the initial schema (in merlin.sql) in sync with the
latest version. This means that an initial install is always the base
schema (version 2, as of this commit - and that shouldn't change), plus
any incremental updates (update-db-XtoX+1.sql).

We also no longer have to update the database version inside the update
scripts - this is now done in install_merlin.sh at the end of each
database update.

In addition, we no longer drop the id column (which is now 8 bytes long
in order to avoid primary key wraparound) in a freakish instance of
megalomania in the install script.

This closes #7994.

Change-Id: I440a26a3df5eb83999b24efe709bd1d00584f11a
Signed-off-by: Anton Lofgren <alofgren@op5.com>