update-db-2to3: add long_output to report_data_extras
The report_data_extras schema should mirror the report_data one (except
for the id).

This patch also fixes a syntax error in the creation of the db version
procedure. Apparently, 'set_initial_db_version' is not the same as
set_initial_db_version (without the quotes). Thus, we've been trying to
drop the wrong procedure. That's bad, because then we can't create it
again, because it already exists.

Change-Id: I0e5a74f497d09ec09aafe5300366efeee82adef8
Signed-off-by: Anton Lofgren <alofgren@op5.com>