ocimp: Make this helper a bit more resilient
Previously we used to bail out hard on failures to write the hash
file, such as when someone had manually done the import as root and
a different user would try it later.

With this patch, we ignore write errors and only care about read
issues, making the program more robust.

Signed-off-by: Andreas Ericsson <ae@op5.se>