Start using STATE_NEGOTIATING
If a master and a poller are of different versions and keep connecting
and disconnecting to each other, a lot of checks will never be executed.
The way around that is to not consider a node connected until we've
successfully negotiated protocol version and node capabilities. This
patch implements exactly that.

Where we previously would not run the checks at all, we now have a
(small) window where it's possible that checks will be run multiple
times from two different nodes instead. That window already exists
anyway, as the time from "try to connect" to "connected" is still a
non-zero number.

Signed-off-by: Andreas Ericsson <ae@op5.se>