mon log: Add --tempdir argument to "import" and "sortmerge"
Since combined logs from several systems can sometimes be extremely
large, people can run out of space in their /tmp directory. This is
especially a problem if they're using ramdisks for /tmp, which is
more or less the norm nowadays.

Allowing them to specify a separate directory means they can sort
even quite huge collections of logfiles, as it makes it possible to
just use a disk partition with enough space to hold the results.

For logs that are too big even for that, some pre-mangling would
be helpful, but for those cases we'll have to deal with it through
support and/or documentation anyways.

Signed-off-by: Andreas Ericsson <ae@op5.se>