import/sql: make insert statement database agnostic
Quoth rsonefors:

Semicolons are not required here, and I know that at least the oracle
libraries we have fails to send queries to the server if they end with a
semicolon.

Backticks for quoting column names are [..] MySQL specific - the
standard way to quote column names are with straight single quotes, "'".

Change-Id: I0e5a74f497d09ec09aafe5300366efeee82adef8
Signed-off-by: Anton Lofgren <alofgren@op5.com>