logrotate: Only start daemons that was running
The previous behaviour was the opposite - shut down monitor and merlin,
and only restart it again if it was shut down previously. Luckily,
without the parent commit, we would never realize that monitor was
running, so we'd always start that again at least, even if we shut
merlin down permanently every sunday.

Change-Id: I488e33526aebe86113946569eedf8b562924436f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>