showlog: Rewrite auth handling
Auth handling is now based on a list of hosts provided on stdin, which
lets ninja grab them from livestatus, rather than forcing showlog to
attempt (and fail) to reach them from an objects.cache file.

While receiving a list of object names on stdin is obviously not as
scalable as one'd like, but there's a test that parses about 10 MB of
hostnames and service descriptions in about a second on my laptop, so if
there's a performance problem, it's not in this part.

This changes the behavior from the old code, in that limited users are
no longer authorized for process and external command messages.

This new API takes care of the merlin-half of #7960.

Change-Id: I2d2ec71f0d4b852fe458e29059845f15cdf890cf
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>