humongos patch: slay oracle support
In issue #8160 we are going to slay oracle support in our codebase

This little patch is what i can find that's relate to OCILIB, ocilib
oracle, Oracle, ORACLE and so on.

Instead of rewrite import.php to work without oracle support we also
throw that and it's deps away since it's no longer maintained and
obsolete since a few versions ago.

GLHF!

Change-Id: I7fa310b181e927aa5e6b41eac2e44eb6b247fe46
Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>