nrpe-merlin.cfg: Remove default command mon_vsz
This command has had its thresholds increased a couple of times
(more specifically 1074ed96fd1 and f337cb515a9b97), which indicates
that something "is rotten in the state of Denmark", as it were.

Since larger installations will require more memory,
the thresholds have been increased to avoid triggering alarms, which
at the same time serves to make the check more and more pointless for
installations of a smaller size. I don't think this speaks in favour
of this being a default service. And so, I've removed it.

Besides, the virtual set size of a process is influenced by more than
just what's been currently heap allocated by the process - which I
gather is what we're really interested in monitoring here (alerting on
memory leaks and whatnot). It also includes (aside from the size of
the mapped binary itself), the size of libraries linked against the
process (resident or not), the size of which could well vary between
different installations.

I have, however, left the mysql_vsz and merlin_vsz commands alone,
since, as far as I know, the problem has not been reported for them
(yet).

Though, I ultimately feel the same about those two commands as well.

This fixes (half of) #7465.

Change-Id: I5463c0d432b763eddd47b0e713ef8950a14875b2
Signed-off-by: Anton Lofgren <alofgren@op5.com>