nagios_command/help: Un-GUI-fy help texts
The help texts for commands have, in many cases, been borrowed from
their respective such in Ninja. This makes sense in some cases, but in
other cases the help texts refer to things such as "check boxes" and
even "options" that is not supported by the command in question, but
rather implemented as a consolidation layer on top of the existing
commands.

Also removed some markup from the help texts, since they make no sense
in this context and otherwise changed some wording.

This is a large, boring patch - go get a beer before starting review.

This fixes #8226.

Change-Id: I85165dd0ac4b7e5ccc47fd5448bb53dd27d71de1
Signed-off-by: Anton Lofgren <alofgren@op5.com>