apps/check: Remove confusing data type transform
What does _dict2Obj do? Well, its true purpose is to typecast all the
values. It also, as a side-effect transforms a dict into an object with
a somewhat dict-like interface. True, this is aesthetics, but I prefer
making the type mangling more obvious, and have the dict-like object be
a dict.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>