apps/test: Put 30 seconds of sleep in random place
In the past, running the "active checks" test took 40 seconds. As soon
as we started tracking passive checks, they started taking 12 seconds.
This is less than one check interval, and made the seemingly unrelated
parent tests fail.

This is, to channel my inner Jack Davenport, without doubt the worst
test suite I've ever heard of. It's an undebuggable, unhelpful piece of
shit that never tells you what went wrong, takes 5 minutes to run not
because it has to but because there are implicit state changes that
probably happens at specified times, and where everything relies
exclusively on what happens before (yet nothing ever properly asserts
that the incoming state is as expected). I've just spent two days
debugging a fucking race in the *test suite*, where nothing gave me a
real indication that that is what was wrong. For God's sake, learn to
write unit tests, or step away from the keyboard!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>