pgroup.c: Refactor pgroup_{host,service}_node into oneliners
This patch introduces a helper which takes the table to look up the
peer group in as an argument. Since that was the only difference
between the two functions earlier they're both made more or less
obsolete, except that the tables are scoped to pgroup.c and therefore
need the wrappers.

Signed-off-by: Andreas Ericsson <ae@op5.se>