(Partially) revert "assignment information: Handle passive checks"
This reverts commit 36bdb7787fb560eccaacf6915be34f4dd3f4269e.

We do retain the separate "responsible" assignment though, as we'll
need it later anyway, and we also keep the check to make sure we
only log once when a check is migrated to a new node (and we don't
spam the logs with "check X moved from A to B" when A and B are
always exactly A and B).

Yes, that message could do with a workover, but *I* know what I
mean.

Signed-off-by: Andreas Ericsson <ae@op5.se>

Conflicts:
	module.c