Add dlist.c, the doubly-linked-list library
We (soon) need a quick way to traverse and remove items from a
list, so a doubly linked list fits our bill. Since it would be
stupid to augment the "linked_item" crap for that, we might as
well put it into a real-ish library so we can reuse it later.

Signed-off-by: Andreas Ericsson <ae@op5.se>