Track the unexpected absence of checkresults
When we're trying to answer the question "is distribution working as
expected?", what we really mean is "have we, in a timely manner, gotten
some sort of result for all the checks we expect to come in?".

With this patch, the module takes care of calculating exactly that
by the relatively simple expedient of adding an expiration event a
short while into the future when Naemon tells us it's time to run a
check. If we fail to get a result for that check (from whatever source)
within a timeframe we deem as reasonable, we'll mark it as expired
and can thus conclude that check distribution/loadbalancing/failover
is flawed in some way.

This paves the way for properly fixing issue #5818, although the patch
almost certainly needs a bit of polish.

Co-authored-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>