module: Tighten up timeouts
We special case daemon startup - hey, the event queue is out of sync
then, after all - and we explain the logic behind checking the last
check result (and move it to where we check for expirations, instead of
putting a technically-incorrect added time).

The over-ambitious slack we added in scheduling can thus go away - we
added extra slack in multiple places until we got zeroes, and it was the
"actually check last check" trickery that actually worked.

Change-Id: I0eed144fae93e335c2e7509c002a1ed83805203f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>