module: Make check unexpiration generic, instant
"Instant" is used in the "user experienced time" sense, not in the
"algorithmic complexity" sense.

Now as soon as we process a check result, we'll make sure said check is
not expired, and remove any currently-scheduled events in the event
queue so they won't even have to check if the event did expire during
the time when we clearly know it didn't.

Change-Id: If503cefb7c3a2c629d9ecf0b7d90592dbca24314
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>