test: Fix test-hooks
Test hooks tested the old status events mechanism for submitting check
results, as well as the new check alert mechanism. This was broken when
merlin stopped supporting the old ways in one of the recent-ancestor
commits. It also wasn't updated to follow along with some of the resent
new dependencies - because manual makefile writing is ridiculously error
prone.

Fix the broken tests so they work, remove the invalid tests, and port
the whole thing to check from the old, awful test_utils.

Change-Id: Ia86308feb13d4eff7343ebafb8fb6d364c13019a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>