mon check: Rewrite dist, orphan around expiration
Expired checks are a better way to detect if a cluster works properly -
not only does it not return incorrect results for the first few minutes,
not only is it not terrible to make it work with passive or infrequent
checks, but it also discovers *which* check is missing from *where*
rather than the old method of just throwing up our hands and saying "the
math doesn't add up".

So this replaces the old mon check distribution, and it also performs
the same task as mon check orphans, so both of them are merged and
replaced with a very short code snippet.

I'd say this resolves #5818, if I wasn't afraid of jinxing it.

Change-Id: Iafb2a0fa4d63aeafc7b478929a450b717d8842b1
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>