mon node status: Stop tracking "assigned"
The reason for #7427 was that mon node status used a very, very
incomplete model of mon check distribution for finding out if everything
ran what it was supposed to or if it needed to redden the numbers.

Remove all that, and instead of exposing the assignment, which is never
correct due to passive and infrequent checks, expose the expired value -
if something is expired, it's bad and red. If not, the check percentages
aren't red.

Change-Id: Ic4f18f13efd4ce046c0de86b1d8f35c123231d0d
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>