daemon: Avoid senseless config push spamming
Without this patch, it's possible to have a single node trigger
contiguous config push events in rapid succession. If the node
triggering these pushes was a poller that we couldn't connect to
and we had peers, we would normally end up restarting our peers
over and over.

This patch adds both belts and suspenders to prevent such config
problem from having a large impact on the network monitoring.

The belts part is that we no longer allow config pushes to be
triggered if the triggering node has triggered one within the
last 30 seconds (not configurable).

Suspenders are provided by allowing each node a maximum of 3
synchronization attempts per connect before we start blocking
them altogether. The 3 attempts can be configured by setting
  max_sync_attempts = X
in the node compound in ones merlin.conf file.

This should fix issue #7492.

Signed-off-by: Andreas Ericsson <ae@op5.se>