apps/test: Fail when we can't start daemons
So... You'd think that a failing test would cause the entire suite to
fail. Not so. The pytap framework that we use for this suite is
fundamentally broken in that it doesn't check children for failures upon
completion. Instead, it somewhat unintuitively checks the current
instance's parents. This is not helpful, since that means you'd have
to keep track of all the information you'd expect a test framework to
keep track of by yourself.  In other words, you'd need to check each
test for failures and then failing them explicitly, since otherwise
you'll only get information about the root instance (which only checks
its own parents, of which there are none). At this point, you might
wonder why you're using a test framework at all, to which I have no
good response.

This is a band-aid. It does not solve the actual problem. And I'm pretty
sure there are more instances of the same symptom in the suite.

Either way, this will at least fail the suite if we can't even start
those bad daemons - which means we won't get builds that haven't had any
tests at all run on them, questionable as they may be.

Change-Id: If30e233d50d5aa233a0f3eb44146b6dd53696508
Signed-off-by: Anton Lofgren <alofgren@op5.com>