Revert "import/sql: make insert statement database agnostic"
This reverts commit 4785664f04d882674668f5b0c0819d260fa8a6e4.

It appears single quoting is not database agnostic at all, just ANSI
SQL. MySQL, in particular, doesn't follow this standard, which causes the
reverted patch to break compatibility with the only database we actually
support.

This is the easy way out. A better solution might be to implement an
actually database-agnostic column-quoting function.

Change-Id: Ib3b66b7af103c304746aa3798a3575c45f55460e