daemon: Force-enable push to be a per-node thing
One can still avoid (and/or override) per-node push commands by
simply specifying object_config { push = no; } inside the node
compound, but global push commands will always be added to each
nodes that don't already have one configured with the name of
the node added to it. It's assumed that the push command will
take care of the argument thus passed and then push the config
for only that node.

This goes a long way towards fixing issue #7492.

Change-Id: If5fbcfa65c70c465dc0e296ef3ed4ab191d9227f
Signed-off-by: Andreas Ericsson <ae@op5.se>