Revert "nrpe-merlin.cfg: Remove default command mon_vsz"
This reverts commit eaea64aeb085deb3fd17650c734831891430aa53.

The revert is done becuase we need to have the old command in the
/etc/nrpe.d/nrpe-merlin.cfg file. The reason for that is becase
the nrpe-merlin.cfg file will be overwritten during an upgrade. If
the command is not in the nrpe-merlin.cfg file old users, that is
using the mon_vsz command, will get a critical alert since the
command has been removed.

This fixes issue: #8490

Signed-off-by: Martin Kamijo <mk@op5.com>