encoding: Let caller supply buffer
We need to deal with message delimiting, the most straightforward way to
do this is to simply prepend each message with its size so that we know
how much to read.

Letting the encode function deal with this implicitly feels a bit icky,
since we nevertheless have to extract that information from the buffer
when decoding ourselves. Thus, I've changed the interface a bit - for
the better, imo.

Change-Id: I5d92b05e3de1d62e88e18668f60ff198aba74841
Signed-off-by: Anton Lofgren <alofgren@op5.com>