Restore old mon check orphan behavior
The goal of the exipiration checking was to do pretty much what mon
check orphan already does, but in a distributed manner. Apparently, the
expiration checking doesn't work quite right when nagios restarts too
much, so go back to doing the stuff that does work.

Restoring behavior takes care of the "more serious" part of #8723.

Change-Id: I830cc001946675c7fa68fb08b86974ac4c8d57a7
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>