daemon: Disconnect when config is out of sync
A node connect is supposed to lead to one of two things:
 * The config is out of sync, so new config is pushed and the other node
   restarted.
 * The config is in sync, so ctrl_active is sent and monitoring begins.

However, in the first case, in the window until the config had been
pushed (which due to max_sync_attempts can be arbitrarily long when one
of them misbehaves) they'll still happily talk to each other and
exchange check results, for items they can't agree on who's supposed to
monitor them.

This resolves bug #8745. It's still possible to get a few error messages
about funky migrations from "internal 'untracked checks'", as it's
possible that one node had already started running the check, and we'll
let it complete rather than discarding a perfectly valid check result -
still, the nodes all log the same thing, and as long as all nodes agree,
we're good.

Change-Id: I1ef199435155938dc8003e4ac93f458dcbb18b9e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>