net: Don't confuse EINPROGRESS and EALREADY
EALREADY means that we tried to connect twice. Stuff be bogus. A warning
message about stuff being "already in progress" is appropriate.

EINPROGRESS means that we tried to connect, in non-blocking mode.
Everything is OK and expected. Saying "already in progress" (which is
pretty much strerror(EALREADY)) is incorrect and confusing, because it
looks like something is wrong.

Change-Id: I80581aab0a80f5aedc5384fe147972d89622ece0
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>