test: Daemons won't connect after we de-sync their config
We have a test which explicitly adds a newline to one node's config, in
an effort to update both its hash and its timestamp, to trigger a config
push. When we have replaced the push command with one that won't
actually push, we shouldn't expect them to reconnect with each other
after that.

This was introduced by e6d1b3b daemon: Disconnect when config is out of
sync

Change-Id: I708c84b184c59ffa56efe5bc239374c0729ad29f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>