module: Handle notification_data events
We have a (currently disabled) hook for it, but no catcher on the
receiving end. This patch adds such a catcher, which really doesn't
do much except incrementing the notification counter for the object
that we're notifying for.

This is necessary to make escalations work properly when nodes go
down after the first notification has been sent and checks (and
thus notifications) are handled by a different node than the
original one at later notifications.

This adds the receiver required to fix issue #8739.

Change-Id: I57b5895f01d55ffd8a9d7007219085e260cb105d
Signed-off-by: Andreas Ericsson <ae@op5.se>