Properly mark host and service notification timings
When a host or service encounters a hard issue and triggers a
notification to be sent, Naemon marks the time when the notification
was sent and calculates the next time we should notify, which also
marks the object for not sending more notifications.

In a distributed environment, those markins weren't set earlier,
but with this patch they are, meaning peers will have the same
idea of when we sent a notification and whether we should send
more or not.

This fixes issue #8668.

Change-Id: Ic865b0850aad3ee4a138a2e15dd328f76a04720b
Signed-off-by: Andreas Ericsson <ae@op5.se>