cfgfile: Make cfgfile's destructor look sane
Now the destructor looks like a destructor, and it no longer risks
triggering a double-free in the - granted, unlikely - event that one
node had a child that didn't know about its parent.

Of course, the absence of tests makes it very hard to know whether this
is correct or not. But it looks nice!

Change-Id: I2a7a17218873d8a58696af3ac494b0669d03e01b
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>