misc: Minor splint/clang-analyzer corrections
A bit of return type corrections to make a splint report merely
unbearable to read, instead of being absolutely impossible (seriously,
we've NEVER EVER looked at a return value for sql_query - just don't
bother...), and a couple of mostly hypothetical junk returns are fixed
as well.

There's a bit of actual bugfixing in there as well, such as using strlen
instead of sizeof for reading the length of a string.

Change-Id: If2d21bf15ee3271a0a6a965c6ed6652cbcb175d2
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>