shared: Kick out safe_free
Checking if a variable is NULL before feeding it to free is useless, as
free does that check anyway.

Supposedly, this breaks compatibility with the compiler on PalmOS. Fuck
PalmOS. The amount of GCCisms and linuxisms in both nagios and merlin is
stunning, and this little trinklet of compatibility just looks silly.

Instead of typedeffing safe_free to free (which is what it was, although
with extra overhead), typedef it to nagios' my_free, which - unlike
safe_free - actually does something worthwhile: it's a macro which sets
the pointer to NULL, so you don't have to remember to do that yourself.

Finally, safe_free makes you safer!

Change-Id: Iccf8f7b8ea5692669d4242cec4bad2aa76d89fd9
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>