module: Carefully transfer the notification_number
This is vital information which was missing: the previous fix for #8739
assumed that the systems were in sync, but as soon as they weren't, they
didn't auto-heal properly.

As most of the merlin protocol is unused padding, anyway, I can just use
one of the pointers to the sender system's memory - which won't be valid
on the receiver anyway - to store the notification number. An int should
be smaller or equal to a uintptr_t everywhere.

Change-Id: I7a06ab2b3644e06ad47f9a16a27c58563398556a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>