daemon: Implement much-logged-about negotiations
When config situations end up being confusing, merlin, like Sophie, has
a choice to make.

It can err on the side of caution, and refuse to talk to the other node
until they have, via the out-of-band ssh mechanism, sorted out their
differences so they can be pals again.

Or it can err on the side of crazy, and decide to ignore the problem and
enter an entirely unhealthy, codependent relationship with the other
node. Because even though both of them know, deep down, that in the long
run, their relationship will only end up hurting both the two of them,
as well as all the people around them who love them and want them to
be strong and monitor all the networks, the alternative of being all
alone seems all too scary.

This commit gives merlin some confidence and courage it so desperately
needs, to be able to stand up to the other nodes and say "no, I won't do
it: I will only be in relationships that are good for me!"

It does so by always sending and requiring the preexisting CTRL_ACTIVE
package during the negotiation phase, instead of assuming it happens
sometime during STATE_CONNECTED. In particular, we now send it *before*
we empty the binlog, not after.

Because the protocol is all too low level, we teach the daemon to
actually disconnect, rather than "marking this node as offline" - this
implies that all of this logic ends up in the daemon, not in the module,
because the module can't disconnect.

This fixes bug #8745, for real this time. This also means that the fix
for bug #7492 is no longer dangerous.

Change-Id: If481a199ddc88d8b63d451d002b289e10c3dda85
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>