test: Fix downtime races
We would register downtimes that were valid for 15 seconds, and then
wait for up to 30 seconds before giving up - this gives you the
impression that none of the downtimes worked, when they did in fact not
only work (most of them, anyway - this test case failed, right?), but
they also expired a long, long time ago.

In addition, we would then wait 10 seconds before hoping that the
downtimes would be gone - which would fail if the test actually *did*
work out within the first 5 seconds.

Change-Id: Ib9f094debcf3f8ec10af3dcfdeb918c83d3a4d0e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>