node: Don't warn about standard cases
Logging is good and all, but logging the fact that you're behaving just
the way you're supposed to is not very helpful, just noisy.

This fixes a bug where Merlin would quickly fill the disk with log
messages about how its cache was full (which is *not* a terrible thing for
a frickin' cache), resulting in an unusable system.

This fixes #8907.

Change-Id: I71a815a9336a430b77b6ef9d2652babdaafd3bdb
Signed-off-by: Anton Lofgren <alofgren@op5.com>