db_updater: Log all UP/OK states as hard
This resolves #8884 by faking hard OK/UP states - soft OK states doesn't
actually exist, yet nagios will log them every once in a while, and then
not log any hard OK states, and then the report data is unreliable.

Technically, this is some seriously next-level bonkers behavior, and I'd
usually just fix it (and thus bug #5199) instead of adding a workaround
around it - however, as pointed out in #5199, this next-level bonkers
behavior is actually intentional and documented. Because of course it
is.

While we try to decide what would be sane and how to fix it, merlin's
data consumers - i.e. ninja's reports - cannot generate reliable
reports. This sucks, and needs to be fixed sooner rather than later,
hence this patch is backported to the 6.3 branch from 7.0.

Once upon a time, this is how merlin actually worked - ticket #18033
actually reports that this exact behavior exists and wants it removed
(even though the bug generated from it, bug #5199, is such an awful bug
report that it has started to mean pretty much the opposite) - so
technically I'm reverting merlin to the earlier, intentional behavior
while awaiting the fix for nagios.

This obviously won't fix the old report data, but an SQL update that
takes care of this would be easy to write - no soft OK states, ever.
Prioritizing logging correct data in the first place is always long-term
preferable to writing workarounds that try to guess what the buggy data
actually means.

Change-Id: Ie20e958751c4fafd50571129434c039f80213d4b
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>