Replace build system with automake
This adds a dependency on the latest naemon HEAD, as I rely on
pkg-config to be able to tell me where I should go look for naemon. On
the other hand, one of the things about pkgconfig is that it lets you
bypass this by simply specifying the naemon_CFLAGS and naemon_LIBS
variables yourself, to skip the process and just use known good values -
hey, it's even documented in the configure script! So that particular
incompatibility can be worked around.

The configure script actually checks if we can find naemon and libdbi
now, instead of just hoping for the best and spewing an ugly error
message at the user like we liked to do in the past.

The unwieldy, handcranked makefile becomes just a few lines of automake,
and dependency tracking is now automatic.

Most of install-merlin.sh has been removed - half of it just parsed
command lines, which the configure script now does. Most of the rest
concerned installing the application itself into god knows where - now,
we default to LSB paths, but as it's automake, the process for
overriding them and get the old paths is predictable and standardize.

This doesn't fix everything - in particular, the database setup is
busted, and will be fixed shortly.

Change-Id: Ib2368e444cbe4b26910f25257baf9c934ed04454
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>