shared: Remove ancient nagios remains
I'm working towards naemon, where NSCORE is dead, so this can go.

As naemon isn't at all compatible with nagios, this means there's a bit
of a limbo in the middle of this patch series - but if you let me, I'll
try to patch it back together again at the end.

Change-Id: I824813dca59bcb19272b43b68f73de4c966e213a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>