makefile: Handle The install-merlin.sh Problem
The problem that I created in the parent commit.

It is handled by installing init-merlin.sh, putting the variables from
the configure script in place (which sucks when you want to change your
password - this could easily be fixed by rewriting the script in python
so it would parse your config, though, but I haven't gotten that far,
yet), but have the users create the database themselves - this way,
we don't need to ask for or remember super-user information.

Change-Id: If374011cfe0b637a749f720c1d044adcefbc2cec
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>